+++ device/src/bt_conn.c
@@ -699,6 +699,7 @@ uint8_t BtConn_UnusedPeripheralConnectionCount() {
     return count;
 }
 
+// Unused in left half
 ATTR_UNUSED static void disconnectOldestHost() {
     uint32_t oldestSwitchover = UINT32_MAX;
     uint8_t oldestPeerId = PeerIdUnknown;