+++ device/src/main.c
@@ -1,5 +1,6 @@
 #include "main.h"
 #include "bt_advertise.h"
+#include "messenger_queue.h"
 #include "nus_client.h"
 #include "nus_server.h"
 #include "bt_manager.h"
@@ -41,6 +42,7 @@
 #include <zephyr/drivers/gpio.h>
 #include "dongle_leds.h"
 #include "usb_protocol_handler.h"
+#include "trace.h"
 
 k_tid_t Main_ThreadId = 0;
 
@@ -72,6 +74,8 @@ static void scheduleNextRun() {
     CurrentTime = k_uptime_get();
     int32_t diff = nextEventTime - CurrentTime;
 
+    Trace('-');
+
     k_sem_take(&mainWakeupSemaphore, K_NO_WAIT);
     bool haveMoreWork = (EventScheduler_Vector & EventVector_UserLogicUpdateMask);
     if (haveMoreWork) {
@@ -90,6 +94,7 @@ static void scheduleNextRun() {
         k_sem_take(&mainWakeupSemaphore, K_FOREVER);
         // k_sleep(K_FOREVER);
     }
+    Trace('+');
 }
 
 //TODO: inline this
@@ -101,6 +106,8 @@ void Main_Wake() {
 int main(void) {
     Main_ThreadId = k_current_get();
     printk("----------\n" DEVICE_NAME " started\n");
+
+    Trace_Init();
 
     {
         flash_area_open(FLASH_AREA_ID(hardware_config_partition), &hardwareConfigArea);