+++ device/src/messenger_queue.c
@@ -6,6 +6,7 @@
 #include <zephyr/kernel.h>
 #include <zephyr/sys/printk.h>
 #include "logger.h"
+#include "trace.h"
 
 uint16_t MessengerQueue_DroppedMessageCount = 0;
 
@@ -112,6 +113,7 @@ uint8_t* MessengerQueue_AllocateMemory() {
     for (uint8_t tries = 0; tries < POOL_SIZE; tries++) {
         POOL_ALLOCATE(regionPool, POOL_SIZE, POOL_REGION_SIZE);
     }
+    Trace_Print();
     LogUOS("Messanger message pool space ran out!\n");
     return blackholeBuffer;
 }
@@ -168,6 +170,16 @@ messenger_queue_record_t MessengerQueue_Take() {
         };
         return res;
     }
+}
+
+ATTR_UNUSED uint8_t MessengerQueue_GetOccupiedCount() {
+    uint8_t count = 0;
+    for (uint8_t i = 0; i < POOL_SIZE; i++) {
+        if (regionPool.segmentTaken[i]) {
+            count++;
+        }
+    }
+    return count;
 }
 
 void MessengerQueue_Init() {