+++ device/src/state_sync.c
@@ -34,7 +34,7 @@
 
 #define WAKE(TID) if (TID != 0) { k_wakeup(TID); }
 
-#define STATE_SYNC_SEND_DELAY 2
+#define STATE_SYNC_SEND_DELAY 1
 
 #define THREAD_STACK_SIZE 2000
 #define THREAD_PRIORITY 5
@@ -58,9 +58,9 @@ uint16_t StateSync_DongleResetCounter = 0;
 static void wake(k_tid_t tid) {
     if (tid != 0) {
         k_wakeup(tid);
-        if (DEBUG_MODE) {
-            printk("StateSync woke up %p\n", tid);
-        }
+        // if (DEBUG_MODE) {
+        //     LogU("StateSync woke up %p\n", tid);
+        // }
     } else {
         printk("Skipping wake up, tid is 0");
     }
@@ -263,13 +263,11 @@ void receiveBacklight(sync_command_backlight_t *buffer) {
     }
 }
 
-static void checkFirmwareVersions(const uhk_module_state_t *moduleState, slot_t slotId) {
-    /*
+void StateSync_CheckFirmwareVersions() {
     #if DEVICE_IS_UHK80_RIGHT
 
-    if (slotId != SlotId_LeftKeyboardHalf) {
-        return;
-    }
+    uint8_t driverId = UhkModuleSlaveDriver_SlotIdToDriverId(SlotId_LeftKeyboardHalf);
+    uhk_module_state_t *moduleState = &UhkModuleStates[driverId];
 
     bool versionsMatch = VERSIONS_EQUAL(moduleState->firmwareVersion, firmwareVersion);
     bool leftChecksumMatches = memcmp(moduleState->firmwareChecksum, DeviceMD5Checksums[DeviceId_Uhk80_Left], MD5_CHECKSUM_LENGTH) == 0;
@@ -284,51 +282,34 @@ static void checkFirmwareVersions(const uhk_module_state_t *moduleState, slot_t 
     }
 
     const char* universal = "Please flash both halves to the same version!";
-    bool fine = true;
-    static bool lastLogWasSuccess = false;
 
     if (!versionsMatch) {
-        fine = false;
-        LogUOS("Error: Left and right keyboard halves have different firmware versions (Left: %u, Right: %u)!\n", moduleState->firmwareVersion, firmwareVersion);
+        LogUOS("Error: Left and right keyboard halves have different firmware versions (Left: %d.%d.%d, Right: %d.%d.%d)!\n",
+            moduleState->firmwareVersion.major, moduleState->firmwareVersion.minor, moduleState->firmwareVersion.patch, firmwareVersion.major, firmwareVersion.minor, firmwareVersion.patch
+        );
     }
     if (!gitTagsMatch) {
-        fine = false;
         LogUOS("Error: Left and right keyboard halves have different git tags (Left: %s, Right: %s)!\n", moduleState->gitTag, gitTag);
     }
     if (!leftChecksumMatches) {
-        fine = false;
         LogUOS("Error: Left checksum differs from the expected! Expected '%s', got '%s'!\n", DeviceMD5Checksums[DeviceId_Uhk80_Left], moduleState->firmwareChecksum);
     }
     if (!versionsMatch || !gitTagsMatch || !leftChecksumMatches) {
-        fine = false;
         LogUOS("    %s", universal);
     }
     if (anyVersionZero) {
-        fine = false;
         LogUOS("Warning: Keyboard halves have zero versions! %s\n", universal);
     }
     if (anyChecksumZero) {
-        fine = false;
         LogUOS("Warning: Keyboard halves have zero checksums! %s\n", universal);
     }
 
-    if (fine && !lastLogWasSuccess) {
-        LogUOS("Left and right keyboard halves have matching firmware versions, git tags and checksums now!\n");
-        lastLogWasSuccess = true;
-    }
-
-    lastLogWasSuccess = fine;
     #endif
-    */
 }
 
 static void checkDongleProtocolVersion() {
-    if (VERSIONS_EQUAL(DongleProtocolVersion, dongleProtocolVersion)) {
-        LogUOS("Dongle and right half run the same dongle protocol version %d.%d.%d\n",
-                DongleProtocolVersion.major, DongleProtocolVersion.minor, DongleProtocolVersion.patch
-        );
-    } else {
-        LogUOS("Dongle and right half run different dongle protocol versios (dongle: %d.%d.%d, right: %d.%d.%d), please upgrade!\n",
+    if (!VERSIONS_EQUAL(DongleProtocolVersion, dongleProtocolVersion)) {
+        LogUOS("Dongle and right half run different dongle protocol versios\n  (dongle: %d.%d.%d, right: %d.%d.%d)\n  please upgrade!\n",
                 DongleProtocolVersion.major, DongleProtocolVersion.minor, DongleProtocolVersion.patch,
                 dongleProtocolVersion.major, dongleProtocolVersion.minor, dongleProtocolVersion.patch
         );
@@ -353,8 +334,6 @@ static void receiveModuleStateData(sync_command_module_state_t *buffer) {
     Utils_SafeStrCopy(moduleState->gitRepo, buffer->gitRepo, MAX_STRING_PROPERTY_LENGTH);
     Utils_SafeStrCopy(moduleState->gitTag, buffer->gitTag, MAX_STRING_PROPERTY_LENGTH);
     memcpy(moduleState->firmwareChecksum, buffer->firmwareChecksum, MD5_CHECKSUM_LENGTH);
-
-    checkFirmwareVersions(moduleState, buffer->slotId);
 
     if (DEVICE_IS_UHK80_RIGHT && leftModuleChanged) {
         EventVector_Set(EventVector_KeymapReloadNeeded);
@@ -399,7 +378,7 @@ static void receiveProperty(device_id_t src, state_sync_prop_id_t propId, const 
     case StateSyncPropertyId_ActiveLayer:
         if (!isLocalUpdate) {
             if (ActiveLayer >= LayerId_Count) {
-                LogUOS("Received invalid active layer %d --- %d %d %d %d %d | %d %d | %d %d\n", ActiveLayer, data[-5], data[-4], data[-3], data[-2], data[-1], data[0], data[1], data[2], data[3]);
+                LogU("Received invalid active layer %d --- %d %d %d %d %d | %d %d | %d %d\n", ActiveLayer, data[-5], data[-4], data[-3], data[-2], data[-1], data[0], data[1], data[2], data[3]);
                 ActiveLayer = LayerId_Base;
             }
             EventVector_Set(EventVector_LedMapUpdateNeeded);
@@ -430,11 +409,10 @@ static void receiveProperty(device_id_t src, state_sync_prop_id_t propId, const 
         // TODO
         break;
     case StateSyncPropertyId_KeyboardLedsState:
-        if (!isLocalUpdate) {
-            WIDGET_REFRESH(&StatusWidget);
-            if (DongleProtocolVersion.major == 0) {
-                LogUOS("Dongle protocol version doesn't seem to have been reported. Is your dongle firmware up to date?\n");
-            }
+        WIDGET_REFRESH(&StatusWidget);
+
+        if (!isLocalUpdate && DongleProtocolVersion.major == 0) {
+            LogUOS("Dongle protocol version doesn't seem to have been reported.\nIs your dongle firmware up to date?\n");
         }
         break;
     case StateSyncPropertyId_ResetRightLeftLink:
@@ -649,6 +627,7 @@ static void prepareData(device_id_t dst, const uint8_t *propDataPtr, state_sync_
     case StateSyncPropertyId_KeyStatesDummy: {
 #if DEVICE_IS_KEYBOARD
         KeyScanner_ResendKeyStates = true;
+        UhkModuleDriver_ResendKeyStates = true;
 #endif
         return;
     }
@@ -686,77 +665,89 @@ static void updateProperty(state_sync_prop_id_t propId) {
     }
 }
 
-#define UPDATE_AND_RETURN_IF_DIRTY(propId)                                                         \
+typedef enum {
+    UpdateResult_AllUpToDate,
+    UpdateResult_UpdatedHighPrio,
+    UpdateResult_UpdatedLowPrio,
+    UpdateResult_UpdatedDelayed,
+} update_result_t;
+
+#define UPDATE_AND_RETURN_IF_DIRTY(propId, res)                                                    \
     if (stateSyncProps[propId].dirtyState != DirtyState_Clean) {                                   \
         updateProperty(propId);                                                                    \
-        return false;                                                                              \
+        return res;                                                                                \
     }
 
-static bool handlePropertyUpdateRightToLeft() {
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ResetRightLeftLink);
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_Config);
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_SwitchTestMode);
+static update_result_t handlePropertyUpdateRightToLeft() {
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ResetRightLeftLink, UpdateResult_UpdatedLowPrio);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_Config, UpdateResult_UpdatedHighPrio);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_SwitchTestMode, UpdateResult_UpdatedHighPrio);
 
     if (KeyBacklightBrightness != 0) {
         // Update relevant data
-        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_PowerMode);
-        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_FunctionalColors);
-        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_LayerActionFirst + ActiveLayer);
-        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ActiveKeymap);
-        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ActiveLayer);
-        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_Backlight);
+        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_PowerMode, UpdateResult_UpdatedHighPrio);
+        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_FunctionalColors, UpdateResult_UpdatedHighPrio);
+        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_LayerActionFirst + ActiveLayer, UpdateResult_UpdatedHighPrio);
+        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ActiveKeymap, UpdateResult_UpdatedHighPrio);
+        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ActiveLayer, UpdateResult_UpdatedHighPrio);
+        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_Backlight, UpdateResult_UpdatedHighPrio);
 
         // Update rest of layers
         state_sync_prop_id_t first = StateSyncPropertyId_LayerActionFirst;
         state_sync_prop_id_t last = StateSyncPropertyId_LayerActionLast;
         for (state_sync_prop_id_t propId = first; propId <= last; propId++) {
-            UPDATE_AND_RETURN_IF_DIRTY(propId);
+            UPDATE_AND_RETURN_IF_DIRTY(propId, UpdateResult_UpdatedLowPrio);
         }
+    } else {
+        UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_Backlight, UpdateResult_UpdatedHighPrio);
     }
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_Backlight);
 
-    return true;
+    return UpdateResult_AllUpToDate;
 }
 
 static bool handlePropertyUpdateLeftToRight() {
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ResetRightLeftLink);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ResetRightLeftLink, UpdateResult_UpdatedLowPrio);
 
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ModuleStateLeftHalf);
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ModuleStateLeftModule);
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_KeyStatesDummy);
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_LeftModuleDisconnected);
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_MergeSensor);
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_Battery);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ModuleStateLeftHalf, UpdateResult_UpdatedHighPrio);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ModuleStateLeftModule, UpdateResult_UpdatedHighPrio);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_KeyStatesDummy, UpdateResult_UpdatedHighPrio);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_LeftModuleDisconnected, UpdateResult_UpdatedHighPrio);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_MergeSensor, UpdateResult_UpdatedHighPrio);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_Battery, UpdateResult_UpdatedHighPrio);
 
-    return true;
+    return UpdateResult_AllUpToDate;
 }
 
 static bool handlePropertyUpdateDongleToRight() {
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ResetRightDongleLink);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ResetRightDongleLink, UpdateResult_UpdatedLowPrio);
 
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_DongleProtocolVersion);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_DongleProtocolVersion, UpdateResult_UpdatedHighPrio);
 
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_KeyboardLedsState);
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_DongleScrollMultipliers);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_KeyboardLedsState, UpdateResult_UpdatedHighPrio);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_DongleScrollMultipliers, UpdateResult_UpdatedHighPrio);
 
-    return true;
+    return UpdateResult_AllUpToDate;
 }
 
 static bool handlePropertyUpdateRightToDongle() {
-    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ResetRightDongleLink);
+    UPDATE_AND_RETURN_IF_DIRTY(StateSyncPropertyId_ResetRightDongleLink, UpdateResult_UpdatedLowPrio);
 
-    return true;
+    return UpdateResult_AllUpToDate;
 }
 
 static void updateLoopRightLeft() {
+    update_result_t res;
+
     if (DEVICE_ID == DeviceId_Uhk80_Left) {
         while (true) {
             bool isConnected = DeviceState_IsDeviceConnected(DeviceId_Uhk80_Right);
             STATE_SYNC_LOG("--- Left to right update loop, connected: %i\n", isConnected);
-            if (!isConnected || handlePropertyUpdateLeftToRight()) {
+
+            if (!isConnected || (res = handlePropertyUpdateLeftToRight()) == UpdateResult_AllUpToDate) {
                 k_sleep(K_FOREVER);
             } else {
-                k_sleep(K_MSEC(STATE_SYNC_SEND_DELAY));
+                uint32_t delay = res == UpdateResult_UpdatedHighPrio ? STATE_SYNC_SEND_DELAY : STATE_SYNC_SEND_DELAY*10;
+                k_sleep(K_MSEC(delay));
             }
         }
     }
@@ -765,10 +756,11 @@ static void updateLoopRightLeft() {
         while (true) {
             bool isConnected = DeviceState_IsDeviceConnected(DeviceId_Uhk80_Left);
             STATE_SYNC_LOG("--- Right to left update loop, connected: %i\n", isConnected);
-            if (!isConnected || handlePropertyUpdateRightToLeft()) {
+            if (!isConnected || (res = handlePropertyUpdateRightToLeft()) == UpdateResult_AllUpToDate) {
                 k_sleep(K_FOREVER);
             } else {
-                k_sleep(K_MSEC(STATE_SYNC_SEND_DELAY));
+                uint32_t delay = res == UpdateResult_UpdatedHighPrio ? STATE_SYNC_SEND_DELAY : STATE_SYNC_SEND_DELAY*10;
+                k_sleep(K_MSEC(delay));
             }
         }
     }
@@ -785,6 +777,8 @@ static void updateStandbys() {
 }
 
 static void updateLoopRightDongle() {
+    update_result_t res;
+
     if (DEVICE_ID == DeviceId_Uhk80_Right) {
         while (true) {
             bool isConnected = DeviceState_IsDeviceConnected(DeviceId_Uhk_Dongle);
@@ -794,10 +788,11 @@ static void updateLoopRightDongle() {
                 updateStandbys();                                                                    \
             }
 
-            if (!isConnected || handlePropertyUpdateRightToDongle()) {
+            if (!isConnected || (res = handlePropertyUpdateRightToDongle()) == UpdateResult_AllUpToDate) {
                 k_sleep(K_FOREVER);
             } else {
-                k_sleep(K_MSEC(STATE_SYNC_SEND_DELAY));
+                uint32_t delay = res == UpdateResult_UpdatedHighPrio ? STATE_SYNC_SEND_DELAY : STATE_SYNC_SEND_DELAY*10;
+                k_sleep(K_MSEC(delay));
             }
         }
     }
@@ -806,10 +801,11 @@ static void updateLoopRightDongle() {
         while (true) {
             bool isConnected = DeviceState_IsDeviceConnected(DeviceId_Uhk80_Right);
             STATE_SYNC_LOG("--- Dongle update loop, connected: %i\n", isConnected);
-            if (!isConnected || DongleStandby || handlePropertyUpdateDongleToRight()) {
+            if (!isConnected || DongleStandby || (res = handlePropertyUpdateDongleToRight()) == UpdateResult_AllUpToDate) {
                 k_sleep(K_FOREVER);
             } else {
-                k_sleep(K_MSEC(STATE_SYNC_SEND_DELAY));
+                uint32_t delay = res == UpdateResult_UpdatedHighPrio ? STATE_SYNC_SEND_DELAY : STATE_SYNC_SEND_DELAY*10;
+                k_sleep(K_MSEC(delay));
             }
         }
     }
@@ -857,6 +853,8 @@ void StateSync_ResetRightLeftLink(bool bidirectional) {
         invalidateProperty(StateSyncPropertyId_Backlight);
         invalidateProperty(StateSyncPropertyId_FunctionalColors);
         invalidateProperty(StateSyncPropertyId_PowerMode);
+        // Wait sufficiently log so the firmware check isnt triggered during firmware upgrade
+        EventScheduler_Schedule(CurrentTime + 60000, EventSchedulerEvent_CheckFwChecksums, "Reset left right link");
     }
     if (DEVICE_ID == DeviceId_Uhk80_Left) {
         invalidateProperty(StateSyncPropertyId_Battery);