+++ right/src/usb_commands/usb_command_get_device_property.c
@@ -81,10 +81,14 @@ void UsbCommand_GetDeviceProperty(const uint8_t *GenericHidOutBuffer, uint8_t *G
             case DEVICE_ID_UHK60V2_RIGHT:
                 if (moduleId == ModuleId_RightKeyboardHalf) {
                     checksum = DeviceMD5Checksums[DEVICE_ID];
-                } else if (moduleId < ModuleId_ModuleCount) {
+                } else if (moduleId == ModuleId_TouchpadRight) {
+                    // return empty buffer
+                    return;
+                } else if (moduleId < ModuleId_AllModuleCount) {
                     checksum = ModuleMD5Checksums[moduleId];
                 } else {
                     SetUsbTxBufferUint8(0, UsbStatusCode_GetDeviceProperty_InvalidArgument);
+                    return;
                 }
                 break;
             case DEVICE_ID_UHK80_LEFT:
@@ -100,8 +104,16 @@ void UsbCommand_GetDeviceProperty(const uint8_t *GenericHidOutBuffer, uint8_t *G
                     case ModuleId_Dongle:
                         checksum = DeviceMD5Checksums[DeviceId_Uhk_Dongle];
                         break;
+                    case ModuleId_TouchpadRight:
+                        // return empty buffer
+                        return;
                     default:
-                        checksum = ModuleMD5Checksums[moduleId];
+                        if (moduleId < ModuleId_AllModuleCount) {
+                            checksum = ModuleMD5Checksums[moduleId];
+                        } else {
+                            SetUsbTxBufferUint8(0, UsbStatusCode_GetDeviceProperty_InvalidArgument);
+                            return;
+                        }
                         break;
                 }
                 break;